Large overhaul of data model.
This adds in basic user roles and ensures email presence over name

Roles are: user, admin, and superadmin. Super admin is reserved for devs
who need global access. All application use cases go through user or
admin.

There are basic policy checks on the create endpoint. Everything else is
open.

The experience management is all done via a single experience api. The
"lobby" concept is just a state of an experience. The front end can
still make lobby pages, or pages to other states, but from a
routing/logic perspective I've tried to simplify this down to a single
entity (experience). For the game management (interactivity) this will
be modelled by states of the experience, which the front-end can consume
and build on top of on any way it needs. This will de couple the API
from the actual user experiences we want to build. It also makes it more
"graphql" like which will ease that migration if we want to go that way.

Application auth is done with passwordless. This renders the default
routes, which look like garbage. I've moved the SPA routing off to a
dedicated controller. This will help in the future for meta tags and
other SPA concerns, and allow us to serve up other content if needed.
E.g the passwordless sign-in pages. This doesn't stop us from moving the
passwordless experience to the SPA, just moves us forward quickly.

Experience Flow.

The other big change here is the formalization of the experience flow.

I've modelled this in a mermaid chart under documentation as it is
actually slightly more complex than we had it modelled.

What I have setup is as follows:
* Step 1 "Join".

This refers to a page that will ask for a "code". This can be directly
navigated to, or redirected to from another source.

I imagine if a QR code is used, or another redirect (e.g a ticketing
system integration gives a link to a user) we don't actually need a user
to input a code. It'll auto-fill and submit, or skip this page and go
straight to step 2.

* Step 2 "Register"

Here is where we collect details and see if a user is able to "register"
for the event.

It asks for an email and optionally a name. The name is in the wrong
place here, as what we really need is a "username" which will be saved
on the experience_participant. Returning users can have it prefilled
with the "most recent" exerpience_participant user name, as an example.
but for now I've just optionally kept the field here.

Here is where we can also perform "authorization" in the future. For
example, if we know a person has paid for a ticket in advance, we should
have knowledge of that and we'll know we can register that email as an
in person experience_participant. Other options here will be things like
a ticketing integration where a user gets a unique code with their
ticket. Register is the verb that encapsulates all of that.

* Step 3 "Experience"

After registration someone is "in the experience". I figure we can model
this like a game in the future with state transitions. For now it just
always returns a "lobby" state and the experience page renders lobby
stuff from the lobby component. Obviously lots of cleanup here but I
wanted to get this in place quickly so people could hack on the frontend
of the experience and start getting patterns they liked in place.

Misc:
* Remove all finger print logic. That was a hack to get session less state
  for the demo. The col still exists as optional as it is a viable future
  state, but not needed now.
* There is no create page, I know this is built on other branches. I
  just have a back end api route for it.
* I'm not using the auth, just generating it, storing it, and passing it
  around.