Datamodel to support experiences.
First pass at modeling how to implement interactivity.

At the top level an experience has a status. This is a mutually
exclusive state to indicate if an experience is drafting, waiting to
start (lobby), live (everything goes here), paused (likely not used but
can be helpful to "turn off" interactivity globally, then two
administrative statuses: finished, and archived.

Each participant now has a role:
* Audience: spectator, default role
* Player: anyone who can interact
* Moderator: not used in this PR but adding it here to show where
  content moderation fits into this model
* Host: The only role that can perform state changes

Experience "Blocks"
Open to a better name on this. State machine nomenclature wasn't helping
out so I went super generic. This is the "unit" that models everything.
All blocks are responsible for:
* Is this drafting, open, closed
* Who can see this: modeled at participant role, segment (custom list),
  specific users
* payload - json column to keep everything we need to model an
  interactive element. We can design schemas on the fly for anything we
  want to experiment with and store it here

Another way to view this. The experience is a basic state machine. It
can have many "things" going on for any combination of people who are
part of the experience.

Things this doesn't model.

There is no priority or exclusiveness explicitly encoded here. Any rule
like "only one poll open at once" is not part of this model. If that
level of "rule" modeling is needed in the future it can go on the
experience in another col. For block specific constraints (e.g 2 max
submissions, or "this role's submissions are worth 2x", that is all in
the payload).

State changes needed for a block. For example, a "response". When we
build elements we'll need to do things like create new tables for
elements.

E.g

```
create_table :experience_poll_submissions, id: :uuid do |t|
  t.references :experience_step, null: false, foreign_key: true, type: :uuid
  t.references :user,            null: false, foreign_key: true, type: :uuid
  t.jsonb :answer, null: false, default: {}
  t.timestamps
end
```