Real time experience updates.
This implements a hybrid approach to continue leveraging the restful
actions participants and admins can make that change the experience
state, while providing real time updates to the participants.

Note: I haven't set up the mange page to have real time updates, we can
do simple polling for now.

Context:
I previously architected a participant's representation of an experience
to be defined by a "key". This is an implicit concept but it was
intentional as it allowed us to reason about a given experience using a
parition key. This model is typically well suited to a stream/channel
approach.

For example, a participant's experience is defined by their:
experience_id:role:segements:participant_id.

Implementation:
Previously, an experience was fetched via a GET request to the
experience show endpoint. This change implements a websocket connection,
via action cable, where a participant subscribes to their stream.

The set of streams is derived from the experience, and each stream is
broadcast to when an update appears. This set can grow large depending
on the number of segments and target users.

A future optimization here is to enqueue broadcasts async. This will
decouple the writes/state changes from the broadcasts, and we can
enforce that N writes doesn't trigger N broadcasts. Furthermore, the
derived streams can each be enqueued independently, so the broadcast
fans out to further optimize.